jobsTodo
The jobs to build can be approximated as a subset of config.jenkins.job and
the contexts found in combiCommitStatus.statuses that are jenkins jobs.

If not rebuilding or gathering all jobs, this subset is either empty or the
main job (if no statuses were found for it). Unless gatherAllJobs, the result
only includes jobs whose most recent status was a failure.